Move Tooling API model checker fixture to appropriate project
